aura-palette-ai/requirements.txt
torch>=1.12
torchvision>=0.13
pandas
Pillow
opencv-python
flask
scikit-learn
tqdm
numpy
joblib
flask-cors
transformers